Complete Phase 2 & 3: Full AgentSessionManager integration with state persistence
- Phase 2: EdgeWorker Integration
  - Integrated AgentSessionManager into handleClaudeMessage() for streaming message capture
  - Added AgentSessionManager initialization for each repository with LinearClient
  - Updated handleClaudeComplete() and handleClaudeError() to mark sessions complete/error
  - All Claude Code SDK messages now automatically tracked in Linear Agent Sessions

- Phase 3: State Management
  - Created proper AgentSession and AgentSessionEntry types in core package
  - Replaced TODO 'any' types with proper TypeScript types
  - AgentSessionManager state now persists across restarts via serializeMappings/restoreMappings
  - Fixed circular dependency by moving types to core package

- Technical improvements:
  - Fixed all TypeScript compilation errors
  - Removed unused LinearDocument import from EdgeWorker
  - Added proper type imports from cyrus-core
  - All 56 tests pass, build completes successfully

The AgentSessionManager now provides comprehensive tracking of Claude Code sessions in Linear with full type safety and state persistence.
